Allow commas in passwords.
Commas are still not supported in usernames or database
names.  However those characters are more likely in
passwords than anywhere else.  There has been a report
of not supporting them in password, but not in the other
locations.  So it was decided to just fix that as the fixes
for all is much more complicated restructure of the configuration.