Update README.md with better information.
Now that we have performance that is much better than before
it can't really be claimed that NMIG is better for certain and
no direct users to use that instead.  It's been moved to an alternative
at the bottom of the page.

Other changes were made to bring the README up to date with many of
the recently merged changes.  These include the ability to have the
application on any computer, not just the server.  The performance
options that may assist you.  The requirement for the PGSQL module
in PHP.